Python: Add `Table` model, save all table references to `metadata.tables`, remove `table_name` and `schema` from metadata.
