+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Unreleased
-- Build Script: Replace version instead of appending to the index file.
+- Add `Table` model and save all table references in `metadata.tables`
 
 
 ## 2024/05/21 v0.0.2
