+++ cratedb_sqlparse_py/tests/test_enricher.py
@@ -1,49 +1,12 @@
 def test_table_metadata():
     from cratedb_sqlparse import sqlparse
-    from cratedb_sqlparse.parser import Metadata
+    from cratedb_sqlparse.models import Metadata
 
     query = "SELECT 1; SELECT 2;"
     stmts = sqlparse(query)
     for stmt in stmts:
         assert hasattr(stmt, "metadata")
         assert isinstance(stmt.metadata, Metadata)
-
-
-def test_table_name_statement():
-    from cratedb_sqlparse import sqlparse
-
-    query = "CREATE TABLE doc.tbl2 (a TEXT)"
-
-    stmts = sqlparse(query)
-    stmt = stmts[0]
-
-    assert stmt.metadata.schema == "doc"
-    assert stmt.metadata.table_name == "tbl2"
-
-
-def test_table_name_statements():
-    from cratedb_sqlparse import sqlparse
-
-    query = """
-    SELECT A,B,C,D,E FROM doc.tbl1;
-    SELECT A,B FROM "doc"."tbl1";
-    SELECT A,B FROM "tbl1";
-    SELECT A,B FROM tbl1;
-    """
-
-    stmts = sqlparse(query=query)
-
-    assert stmts[0].metadata.schema == "doc"
-    assert stmts[0].metadata.table_name == "tbl1"
-
-    assert stmts[1].metadata.schema == "doc"
-    assert stmts[1].metadata.table_name == "tbl1"
-
-    assert stmts[2].metadata.schema is None
-    assert stmts[2].metadata.table_name == "tbl1"
-
-    assert stmts[3].metadata.schema is None
-    assert stmts[3].metadata.table_name == "tbl1"
 
 
 def test_table_with_properties():
@@ -68,3 +31,20 @@ def test_with_with_parameterized_properties():
     # Has all the keys.
     assert stmt.metadata.with_properties == expected
     assert stmt.metadata.parameterized_properties == expected
+
+
+def test_table_names():
+    from cratedb_sqlparse import sqlparse
+
+    query = "SELECT _ FROM a.b, d"
+
+    stmt = sqlparse(query)[0]
+
+    assert len(stmt.metadata.tables) == 2
+    assert stmt.metadata.tables[0].schema == "a"
+    assert stmt.metadata.tables[0].name == "b"
+    assert stmt.metadata.tables[0].fqn == '"a"."b"'
+
+    assert stmt.metadata.tables[1].schema is None
+    assert stmt.metadata.tables[1].name == "d"
+    assert stmt.metadata.tables[1].fqn == '"d"'
